Add FrogBinding
The FrogBinding connects provisioned metadata and secrets into a
PodSpecable resource's container following the CNB Bindings spec.

To summaries the CNB Binding spec:
- a common directory is defined by the `CNB_BINDINGS` envvar
- each binding is a directory inside the common directory
- each binding has a `metadata` directory with non-sensitive values
- each binding may have a `secret` directory with sensitive values

In k8s we represent metadata values with a ConfigMap and the secret
values with a Secret for each binding.

The FrogBinding resource contains a `subject` that defines the
resource(s) where the bound values should be injected, and `providers`
which is a set of provisioned service credentials to inject into the
subject.